+++ packages/utils/src/index.js
@@ -221,3 +221,12 @@ export const versionGreaterOrEqualTo = (comp, base) => {
   const v2 = parseVersion(base);
   return semver.gte(v1, v2);
 };
+
+export const upcase = str => str.substr(0, 1).toUpperCase() + str.substr(1);
+
+// Iteratively execute a callback against each item in an array
+export const asyncForEach = async (array, callback) => {
+  for (let index = 0; index < array.length; index++) {
+    await callback(array[index], index, array);
+  }
+};