+++ packages/app-nuxt/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @keystonejs/app-nuxt
 
+## 5.1.2
+
+### Patch Changes
+
+- [`5ba330b8`](https://github.com/keystonejs/keystone/commit/5ba330b8b2609ea0033a636daf9a215a5a192c20) [#2487](https://github.com/keystonejs/keystone/pull/2487) Thanks [@Noviny](https://github.com/Noviny)! - Small changes to package.json (mostly adding a repository field)
+
 ## 5.1.1
 
 ### Patch Changes
@@ -16,7 +22,7 @@
 
 ### Major Changes
 
-- [`7b4ed362`](https://github.com/keystonejs/keystone-5/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310) [#1821](https://github.com/keystonejs/keystone-5/pull/1821) Thanks [@jesstelford](https://github.com/jesstelford)! - Release @keystonejs/\* packages (つ＾ ◡ ＾)つ
+- [`7b4ed362`](https://github.com/keystonejs/keystone/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310) [#1821](https://github.com/keystonejs/keystone/pull/1821) Thanks [@jesstelford](https://github.com/jesstelford)! - Release @keystonejs/\* packages (つ＾ ◡ ＾)つ
 
   - This is the first release of `@keystonejs/*` packages (previously `@keystone-alpha/*`).
   - All packages in the `@keystone-alpha` namespace are now available in the `@keystonejs` namespace, starting at version `5.0.0`.
@@ -28,4 +34,4 @@
 
 ### Major Changes
 
-- [4b3bb060](https://github.com/keystonejs/keystone-5/commit/4b3bb060): Adding new "app-nuxt" package in @keystone-alpha
+- [4b3bb060](https://github.com/keystonejs/keystone/commit/4b3bb060): Adding new "app-nuxt" package in @keystone-alpha