+++ packages/arch/docs/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @arch-ui/docs
 
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [[`0de5f232`](https://github.com/keystonejs/keystone/commit/0de5f2321ef8f9fe6dd247c3201372a4156e61e9)]:
+  - @arch-ui/theme@0.0.10
+  - @arch-ui/button@0.0.16
+  - @arch-ui/input@0.1.7
+  - @arch-ui/layout@0.2.11
+  - @arch-ui/pill@0.1.13
+  - @arch-ui/alert@0.0.15
+  - @arch-ui/badge@0.0.14
+  - @arch-ui/controls@0.1.6
+  - @arch-ui/dialog@0.0.18
+  - @arch-ui/dropdown@0.0.16
+  - @arch-ui/fields@2.1.4
+  - @arch-ui/loading@0.0.15
+  - @arch-ui/lozenge@0.0.14
+  - @arch-ui/navbar@0.1.8
+  - @arch-ui/pagination@0.0.16
+  - @arch-ui/popout@0.0.17
+  - @arch-ui/select@0.1.6
+  - @arch-ui/tooltip@0.1.9
+  - @arch-ui/typography@0.0.15
+
+## 1.1.14
+
+### Patch Changes
+
+- [`6b353eff`](https://github.com/keystonejs/keystone/commit/6b353effc8b617137a3978b2c845e01403889722) Thanks [@timleslie](https://github.com/timleslie)! - Upgraded React to 16.13.0.
+
+* [`5ba330b8`](https://github.com/keystonejs/keystone/commit/5ba330b8b2609ea0033a636daf9a215a5a192c20) [#2487](https://github.com/keystonejs/keystone/pull/2487) Thanks [@Noviny](https://github.com/Noviny)! - Small changes to package.json (mostly adding a repository field)
+
+* Updated dependencies [[`6b353eff`](https://github.com/keystonejs/keystone/commit/6b353effc8b617137a3978b2c845e01403889722), [`5ba330b8`](https://github.com/keystonejs/keystone/commit/5ba330b8b2609ea0033a636daf9a215a5a192c20), [`9d11d7ce`](https://github.com/keystonejs/keystone/commit/9d11d7cea6f31eeceb7326d86460a3f0e25ad01d)]:
+  - @arch-ui/alert@0.0.14
+  - @arch-ui/badge@0.0.13
+  - @arch-ui/button@0.0.15
+  - @arch-ui/controls@0.1.5
+  - @arch-ui/dialog@0.0.17
+  - @arch-ui/dropdown@0.0.15
+  - @arch-ui/fields@2.1.3
+  - @arch-ui/icons@0.0.10
+  - @arch-ui/input@0.1.6
+  - @arch-ui/layout@0.2.10
+  - @arch-ui/loading@0.0.14
+  - @arch-ui/lozenge@0.0.13
+  - @arch-ui/navbar@0.1.7
+  - @arch-ui/pagination@0.0.15
+  - @arch-ui/pill@0.1.12
+  - @arch-ui/popout@0.0.16
+  - @arch-ui/select@0.1.5
+  - @arch-ui/tooltip@0.1.8
+  - @arch-ui/typography@0.0.14
+  - @arch-ui/theme@0.0.9
+
 ## 1.1.13
 
 ### Patch Changes
@@ -157,41 +212,41 @@
 
 ## 1.1.5
 
-- Updated dependencies [464d7579](https://github.com/keystonejs/keystone-5/commit/464d7579):
+- Updated dependencies [464d7579](https://github.com/keystonejs/keystone/commit/464d7579):
   - @arch-ui/select@0.1.0
 
 ## 1.1.4
 
-- Updated dependencies [7689753c](https://github.com/keystonejs/keystone-5/commit/7689753c):
+- Updated dependencies [7689753c](https://github.com/keystonejs/keystone/commit/7689753c):
   - @arch-ui/controls@0.0.10
   - @arch-ui/input@0.1.0
 
 ## 1.1.3
 
-- Updated dependencies [957a40d9](https://github.com/keystonejs/keystone-5/commit/957a40d9):
+- Updated dependencies [957a40d9](https://github.com/keystonejs/keystone/commit/957a40d9):
   - @arch-ui/select@0.0.8
 
 ## 1.1.2
 
 ### Patch Changes
 
-- [42c3fbc9](https://github.com/keystonejs/keystone-5/commit/42c3fbc9): Upgrade emotion to 10.0.14
-- [42c3fbc9](https://github.com/keystonejs/keystone-5/commit/42c3fbc9): Update gatsby to 2.13.25
+- [42c3fbc9](https://github.com/keystonejs/keystone/commit/42c3fbc9): Upgrade emotion to 10.0.14
+- [42c3fbc9](https://github.com/keystonejs/keystone/commit/42c3fbc9): Update gatsby to 2.13.25
 
 ## 1.1.1
 
-- Updated dependencies [3544f852](https://github.com/keystonejs/keystone-5/commit/3544f852):
+- Updated dependencies [3544f852](https://github.com/keystonejs/keystone/commit/3544f852):
   - @arch-ui/tooltip@0.1.0
 
 ## 1.1.0
 
 ### Minor Changes
 
-- [9f4c2878](https://github.com/keystonejs/keystone-5/commit/9f4c2878):
+- [9f4c2878](https://github.com/keystonejs/keystone/commit/9f4c2878):
 
   Added 'new project', 'adding lists' and 'relationships' guides.
 
 ## 1.0.1
 
-- Updated dependencies [c2dc6eb3](https://github.com/keystonejs/keystone-5/commit/c2dc6eb3):
+- Updated dependencies [c2dc6eb3](https://github.com/keystonejs/keystone/commit/c2dc6eb3):
   - @arch-ui/navbar@0.1.0